Fix subfield code storage
Codes can be different from one instance of a tag to the next, so we
can't only assign the last set of codes seen, we have to add all the
codes, and remove duplicates (to spec).

For other data, we can push (`<<`) data into the array, but here the
data being pushed is always an array, so we must concat, or flatten the
result.